Create Classification-Regression Tree
Created CARTS and compared that against benchmark performance as well as a Linear Regression Model to predict the final sale price of a car if it were to go on auction. I was seeking to leverage this knowledge to place competitive bets and gain a profit on the margin.  Also controlled for overfitting by using both forward "StoppingRules" and backward "Prune" methods to help achieve a balance between model complexity (in this case #of decision splits) and error rate.