Create Neural Networks
Created and tried to understand the operation of Neural Nets. Compared its performance against a Logistic Regression Model and the benchmark error rate.  Plotted Neural Network to get a sense of the hidden layers and figure out what variables/inputs are contributing to final output (which in this case was whether or not a particular marketing campaign was going to be effective at a bank).